contribution being created. updated integration test

+++ src/main/java/com/example/jackpot_service/jackpot/service/FixedJackpotContributionCalculation.java
@@ -22,7 +22,7 @@ public class FixedJackpotContributionCalculation implements JackpotContribCalcul
      * @return The calculated jackpot contribution, which is 1% of the bet amount.
      */
     @Override
-    public int calculateContribution(int betAmount, BigDecimal jackpotPoolAmount) {
+    public int calculateContribution(int betAmount, int jackpotPoolAmount) {
         return BigDecimal.valueOf(betAmount).multiply(FIXED_CONTRIBUTION_PERCENTAGE).setScale(0, RoundingMode.HALF_UP).intValue();
     }
